# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing government spending documents across Federal, State, and Municipal channels\nUses Hyperledger Fabric blockchain for immutable, transparent spending records",
        "title": "Government Spending Blockchain API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/api/anchors/verify": {
            "post": {
                "description": "Verify that two documents are properly linked across channels using cryptographic hashes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Verification"
                ],
                "summary": "Verify cross-channel link",
                "parameters": [
                    {
                        "description": "Source and target document info",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VerifyAnchorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AnchorVerification"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/transfers/initiate": {
            "post": {
                "description": "Start an inter-government transfer (e.g., Federal → State). Creates document with cryptographic hash.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transfers"
                ],
                "summary": "Initiate cross-channel transfer",
                "parameters": [
                    {
                        "description": "Transfer details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.InitiateTransferRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.TransferResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/{channel}/document-types": {
            "get": {
                "description": "Get all document types for a channel",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Document Types"
                ],
                "summary": "List document types",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel (union, state, region)",
                        "name": "channel",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.DocumentType"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new document type template with required/optional fields",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Document Types"
                ],
                "summary": "Register document type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel (union, state, region)",
                        "name": "channel",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Document type data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateDocumentTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.IDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/{channel}/document-types/{typeId}": {
            "get": {
                "description": "Get specific document type by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Document Types"
                ],
                "summary": "Get document type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel (union, state, region)",
                        "name": "channel",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Document type ID",
                        "name": "typeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DocumentType"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Mark document type as inactive (no new documents allowed)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Document Types"
                ],
                "summary": "Deactivate document type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel (union, state, region)",
                        "name": "channel",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Document type ID",
                        "name": "typeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/{channel}/documents/{docId}": {
            "get": {
                "description": "Get specific document by ID with all details",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Get document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel (union, state, region)",
                        "name": "channel",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "docId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Document"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/{channel}/documents/{docId}/history": {
            "get": {
                "description": "Get complete transaction history for a document (all versions)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Get document history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel (union, state, region)",
                        "name": "channel",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "docId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.HistoryEntry"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/{channel}/documents/{docId}/invalidate": {
            "post": {
                "description": "Mark document as invalid (immutable - never deleted). Used for error correction.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Invalidate document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel (union, state, region)",
                        "name": "channel",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "docId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Invalidation reason and correction doc",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.InvalidateDocumentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Only creating organization can invalidate",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.AnchorVerification": {
            "type": "object",
            "properties": {
                "amountMatch": {
                    "type": "boolean"
                },
                "channelMatch": {
                    "type": "boolean"
                },
                "hashMatch": {
                    "type": "boolean"
                },
                "idMatch": {
                    "type": "boolean"
                },
                "isValid": {
                    "type": "boolean"
                },
                "mismatchReason": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sourceAmount": {
                    "type": "number"
                },
                "sourceChannel": {
                    "type": "string"
                },
                "sourceCurrency": {
                    "type": "string"
                },
                "sourceDocId": {
                    "type": "string"
                },
                "sourceHash": {
                    "type": "string"
                },
                "status": {
                    "description": "\"VERIFIED\" or \"MISMATCH\"",
                    "type": "string"
                },
                "targetAmount": {
                    "type": "number"
                },
                "targetChannel": {
                    "type": "string"
                },
                "targetCurrency": {
                    "type": "string"
                },
                "targetDocId": {
                    "type": "string"
                },
                "targetHash": {
                    "type": "string"
                }
            }
        },
...
        "models.IDResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "models.VerifyAnchorRequest": {
            "type": "object",
            "required": [
                "sourceChannel",
                "sourceDocId",
                "targetChannel",
                "targetDocId"
            ],
            "properties": {
                "sourceChannel": {
                    "type": "string"
                },
                "sourceDocId": {
                    "type": "string"
                },
                "targetChannel": {
                    "type": "string"
                },
                "targetDocId": {
                    "type": "string"
                }
            }
        }
    },
...
}